refactor: rename package to `orchestration` and update references in SDK and examples
