+++ orchestration/cluster_test.go
@@ -1,4 +1,4 @@
-package camunda_test
+package orchestration_test
 
 import (
 	"context"
@@ -8,8 +8,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/sijoma/camunda-go-sdk"
 	"github.com/sijoma/camunda-go-sdk/internal"
+	"github.com/sijoma/camunda-go-sdk/orchestration"
 )
 
 func TestWithCamunda(t *testing.T) {
@@ -22,8 +22,8 @@ func TestWithCamunda(t *testing.T) {
 	camundaURL, err := suite.CamundaEndpoint()
 	require.NoError(t, err)
 	baseURL, _ := url.Parse(camundaURL)
-	c8, err := camunda.NewOrchestrationClusterClient(
-		camunda.WithBaseURL(*baseURL),
+	c8, err := orchestration.NewClient(
+		orchestration.WithBaseURL(*baseURL),
 	)
 	require.NoError(t, err)
